+++ src/types/index.ts
@@ -21,3 +21,6 @@ export * from './cpe'
 export * from './cwe'
 export * from './integer'
 export * from './mimeType'
+
+// yes, this is node-specific, but who cares - these are just types
+export * from './nodePackageJson'